+++ test/solc/CommandLineParser.cpp
@@ -198,7 +198,9 @@ BOOST_AUTO_TEST_CASE(cli_mode_options)
 			true, true, true, true, true,
 			true, true, true, true, true,
 			true, true, true, true, true,
+			true,
 		};
+		expectedOptions.compiler.outputs.ewasmIR = false;
 		expectedOptions.compiler.estimateGas = true;
 		expectedOptions.compiler.combinedJsonRequests = {
 			true, true, true, true, true,
@@ -297,6 +299,11 @@ BOOST_AUTO_TEST_CASE(assembly_mode_options)
 				"underflow,"
 				"divByZero",
 			"--model-checker-timeout=5",   // Ignored in assembly mode
+			"--asm",
+			"--bin",
+			"--ir-optimized",
+			"--ewasm",
+			"--ewasm-ir",
 		};
 		commandLine += assemblyOptions;
 		if (expectedLanguage == AssemblyStack::Language::StrictAssembly || expectedLanguage == AssemblyStack::Language::Ewasm)
@@ -333,6 +340,11 @@ BOOST_AUTO_TEST_CASE(assembly_mode_options)
 		};
 		expectedOptions.formatting.coloredOutput = false;
 		expectedOptions.formatting.withErrorIds = true;
+		expectedOptions.compiler.outputs.asm_ = true;
+		expectedOptions.compiler.outputs.binary = true;
+		expectedOptions.compiler.outputs.irOptimized = true;
+		expectedOptions.compiler.outputs.ewasm = true;
+		expectedOptions.compiler.outputs.ewasmIR = true;
 		if (expectedLanguage == AssemblyStack::Language::StrictAssembly || expectedLanguage == AssemblyStack::Language::Ewasm)
 		{
 			expectedOptions.optimizer.enabled = true;